Rename static.lib to default-values
