+++ tests/support/data/licence.data.js
@@ -1,5 +1,5 @@
 import { generateUUID } from '../helpers/generate-uuid.js'
-import { regionCode } from '../static.lib.js'
+import { regionCode } from '../default-values.js'
 
 export default function (licenceRef, companyData) {
   const licenceDocumentHeaderId = generateUUID()
