+++ tests/support/scenarios/registered-licence-with-open-return-log-for-first-period.scenario.js
@@ -1,7 +1,7 @@
 import primaryUserData from '../data/primary-user.data.js'
 import unregisteredLicenceWithOpenReturnLogForFirstPeriod from './unregistered-licence-with-open-return-log-for-first-period.scenario.js'
 import { mergeByKey } from '../helpers/scenario.helpers.js'
-import { externalUserEmail } from '../static.lib.js'
+import { externalUserEmail } from '../default-values.js'
 
 export const title = 'Registered licence with open return log (first period)'
 export const description = 'Registered licence with an open return log for the first return period with no due date set'
